# apps/api/src/openapi/examples/example-response.schemas.ts
import { ApiProperty } from '@nestjs/swagger';

export class ExampleItemSchema {
  @ApiProperty({ format: 'uuid' })
  id!: string;

  @ApiProperty()
  title!: string;

  @ApiProperty()
  body!: string;

  @ApiProperty({ format: 'date-time' })
  createdAt!: string;

  @ApiProperty({ format: 'date-time' })
  updatedAt!: string;
}

export class ListExamplesResponseSchema {
  @ApiProperty({ type: [ExampleItemSchema] })
  items!: ExampleItemSchema[];

  @ApiProperty({ minimum: 1 })
  page!: number;

  @ApiProperty({ minimum: 1 })
  limit!: number;

  @ApiProperty({ minimum: 0 })
  total!: number;
}