# docs/openapi/api-v1.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/rbac/_probe/admin": {
      "get": {
        "operationId": "RbacProbeController_adminProbe_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RbacProbeOkResponseSchema"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "RBAC probe: admin role required",
        "tags": ["rbac"]
      }
    },
    "/api/v1/rbac/_probe/posts-write": {
      "get": {
        "operationId": "RbacProbeController_postsWriteProbe_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RbacProbeOkResponseSchema"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "RBAC probe: posts:write permission required",
        "tags": ["rbac"]
      }
    },
...
    "/api/v1/examples": {
      "post": {
        "operationId": "ExamplesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExampleBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleItemSchema"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "summary": "Create example item",
        "tags": ["examples"]
      },
      "get": {
        "operationId": "ExamplesController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListExamplesResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "summary": "List example items (paginated)",
        "tags": ["examples"]
      }
    },
    "/api/v1/examples/{id}": {
      "get": {
        "operationId": "ExamplesController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleItemSchema"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "summary": "Get example item by id",
        "tags": ["examples"]
      },
      "patch": {
        "operationId": "ExamplesController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExampleBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleItemSchema"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "summary": "Update example item",
        "tags": ["examples"]
      },
      "delete": {
        "operationId": "ExamplesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsBodySchema"
                }
              }
            }
          }
        },
        "summary": "Delete example item",
        "tags": ["examples"]
      }
    }
  },
  "info": {
    "title": "Blog API",
    "description": "Blog/CMS HTTP API. Versioned routes live under `/api/v1`. Errors use `application/problem+json`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token from login or refresh"
      }
    },
    "schemas": {
      "ProblemDetailsBodySchema": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "https://blog.dev/problems/validation-failed"
          },
          "title": {
            "type": "string",
            "example": "Validation Failed"
          },
          "status": {
            "type": "number",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Request body failed validation."
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/auth/login"
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_FAILED"
          },
          "details": {
            "type": "object",
            "additionalProperties": true,
            "example": {
              "fields": [
                {
                  "path": "email",
                  "message": "email must be an email"
                }
              ]
            }
          }
        },
        "required": ["type", "title", "status", "detail", "code"]
      },
      "RbacProbeOkResponseSchema": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [true]
          }
        },
        "required": ["ok"]
      },
...
      "CreateExampleBodyDto": {
        "type": "object",
        "properties": {}
      },
      "ExampleItemSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "title", "body", "createdAt", "updatedAt"]
      },
      "ListExamplesResponseSchema": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExampleItemSchema"
            }
          },
          "page": {
            "type": "number",
            "minimum": 1
          },
          "limit": {
            "type": "number",
            "minimum": 1
          },
          "total": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": ["items", "page", "limit", "total"]
      },
      "UpdateExampleBodyDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}